pyproject.toml
[project]
name = "fraud-detection-graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic>=1.4.3",
    "langchain-aws>=1.4.5",
    "langchain-neo4j>=0.9.0",
    "langgraph>=1.1.10",
    "neo4j",
    "pydantic>=2.13.3",
    "python-dotenv",
]
